places searched will appear in docs(places). /api/routes returns hard coded routes matrix with 1 origin and 2 destinations.
